Split ApiOperations into LorreOperations and ConseilOperations
