+++ conseil-common/src/main/scala/tech/cryptonomic/conseil/common/tezos/TezosDatabaseOperations.scala
@@ -819,7 +819,7 @@ object TezosDatabaseOperations extends LazyLogging {
     Tables.Operations.findBy(_.operationGroupHash)
 
   /** Computes the max level of blocks or [[defaultBlockLevel]] if no block exists */
-  private[tezos] def fetchMaxBlockLevel: DBIO[Int] =
+  def fetchMaxBlockLevel: DBIO[Int] =
     Tables.Blocks
       .map(_.level)
       .max
